boxel-cli/software-factory: address PR #5246 review
- Extract the reusable QUnit engine + types into src/lib/test-engine.ts;
  commands/test.ts is now just the commander wiring, and api.ts imports the
  engine from the lib module so library consumers no longer load command-only
  deps (commander / cliLog / colors).
- boxel-cli: label only the run-end waitForFunction timeout (rethrow other
  Playwright timeouts untouched, with the original as `cause`); document
  `authorization` as a verbatim HTTP Authorization header value, not a bare JWT.
- software-factory: normalize the realm URL with ensureTrailingSlash before
  fetching the token and running (token cache / private-realm auth); drop the
  `as QunitResults` cast so the compiler enforces structural compatibility;
  match only the run-end timeout marker and attach the original error as `cause`.
